added 'projects', removed 'project'
added tasks and project in database
