+++ src/project/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Project, Task
+# from .models import Project
 # Register your models here.
-admin.site.register(Project)
-admin.site.register(Task)
+# admin.site.register(Project)
+# admin.site.register(Task)
