+++ src/project/forms.py
@@ -1,0 +1,16 @@
+from django import forms
+
+# from .models import Project
+
+# class ProjectForm(forms.ModelForm):
+
+#     class Meta:
+#         model = Project
+#         fields = ('projectName', 'teamMembers',)
+
+
+# class TaskForm(forms.ModelForm):
+
+# 	class Meta:
+# 		model = Task
+# 		fields = ('taskName','projectID')
