+++ src/projects/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Project,Task
+# Register your models here.
+admin.site.register(Project)
+admin.site.register(Task)
