+++ src/projects/apps.py
@@ -1,0 +1,7 @@
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    name = 'projects'
