+++ src/together/urls.py
@@ -26,7 +26,7 @@ urlpatterns = [
     url(r'^contact/$', 'users.views.contact', name='contact'),
     url(r'^profile/$', 'users.views.myprofile', name='profile'),
     url(r'^profile/edit/$', 'users.views.editprofile', name='editprofile'),
-    url(r'^projects/$', 'project.views.myprojects', name='projects'),
+    url(r'^projects/$', 'projects.views.myprojects', name='projects'),
     url(r'^discussions/$', 'discussion.views.discussions', name='discussions'),
     url(r'^discussions/post_list', 'discussion.views.post_list', name='post_list'),
     url(r'^discussions/post/(?P<pk>\d+)/$', 'discussion.views.post_detail', name='post_detail'),
